==> notebooks/3_pitchtierToJSON.Rmd <==
---
title: "PitchTier to JSON"
output:
---

Utility 

### Convert PitchTier to JSON - One approach

Reads all the .PitchTier files in $path$ and converts them to JSON files.
Using jsonlite package. Array of points.

```{r}
library("rPraat")
library("jsonlite")
library("stringr")
source("../utils/hertzToST.R")

# Lowest pitch
ref = 60
# Relative path to where the files are
path="../data/pitch-tier-tests/"

setwd(path)
filenames <- dir(pattern="\\.PitchTier$")
setwd("../../notebooks/")
sapply(filenames, pitchTierToJSON)

# Function to turn Praat's PitchTier into JSON file
# Hertz from PitchTier is converted to Semitones, centered on ref pitch
pitchTierToJSON <- function(name) {
  #filename <- paste0(path, name)
  #cat(getwd(), filename)
  # Load PitchTier object using rPraat
  cat("wd ", getwd(), "\n")
  cat("name ", name, "\n")
  my.pt = pt.read(paste0(path, name))
  #my.pt = pt.read(name)
  # Save the time and frequency columns as dataframe
  # divide time by max time
  pt.df <- data.frame(scrub=my.pt$t/my.pt$tmax, f=hertz_to_semitones(my.pt$f, ref))  
  
  # Add a point to the beginning: t = 0 + 0.01 (offset by a little), f = first f
  pt.df <- rbind(c(0.01, pt.df[1,2]), pt.df)
  
  # Add last point: t = tmax - 0.01 (offset by a little), f = last f
  pt.df[nrow(pt.df) + 1,] = c(0.98, pt.df[nrow(pt.df),2])
  
  # Make the dataframe into a JSON object
  my.JSON <- toJSON(pt.df)
  
  # Get the core name for the output filename
  out_name <- str_split(name, ".PitchTier")[[1]][1]
  
  # Write JSON object into the same directory as the original file
  write(my.JSON, paste0(path,out_name, ".json"))
}

```

### Another version
With reference frequency. Need to use $toJSON()$ from the rjson package rather than the jsonlite package in order to put this extra info about ref frequency in JSON file.

```{r}
library("stringr")
library("rPraat")
library("rjson")
source("../utils/hertzToST.R")

# Lowest pitch
ref = 130
# Relative path to where the files are
path="../data/pitch-tier-tests/"

setwd(path)
filenames <- dir(pattern="\\.PitchTier$")
setwd("../../notebooks/")
sapply(filenames, pitchTierToJSON2)

# Function to turn Praat's PitchTier into JSON file
# Hertz from PitchTier is converted to Semitones, centered on ref pitch
pitchTierToJSON2 <- function(name) {
  #filename <- paste0(path, name)
  #cat(getwd(), filename)
  # Load PitchTier object using rPraat
  cat("wd ", getwd(), "\n")
  cat("name ", name, "\n")
  my.pt = pt.read(paste0(path, name))
  #my.pt = pt.read(name)
  # Save the time and frequency columns as dataframe
  # divide time by max time
  pt.df <- data.frame(scrub=my.pt$t/my.pt$tmax, f=hertz_to_semitones(my.pt$f, ref))  
  
  # Add a point to the beginning: t = 0 + 0.01 (offset by a little), f = first f
  pt.df <- rbind(c(0.01, pt.df[1,2]), pt.df)
  
  # Add last point: t = tmax - 0.01 (offset by a little), f = last f
  pt.df[nrow(pt.df) + 1,] = c(0.98, pt.df[nrow(pt.df),2])
  
  # Get the core name for the output filename
  out_name <- str_split(name, ".PitchTier")[[1]][1]
  
  # EXTRA STEP for this version, make a list
  pt.list <- list(points = pt.df, reference=ref, name=out_name)
  
  # Make the dataframe into a JSON object
  my.JSON <- toJSON(pt.list)
  
  # Write JSON object into the same directory as the original file
  write(my.JSON, paste0(path,out_name, ".json"))
}

```
