==> notebooks/5_gestToReprosody.Rmd <==
---
title: "Offline Resynthesis from Gepeto Saved Gesture"
output: html_notebook
---

Takes saved gesture from Gepeto App (.gest) and converts it to a text file that can be used by Grégoire Locqueville's reprosody program to generate resynthesized file. Files for reprosody take the following format:
t_out t_in f. t_out specifies timestamp in seconds in output file, t_in specifies timestamp in seconds of original file. f specifies the frequency we want in hertz. 

This script can be used to study students' gesture inputs in a perceptual test

## Potential Improvements

Currently, Gepeto app does not store the length of sound files, so I need to get the length of the original file separately. It would be helpful to specify the scrub time (0 to 1) in the original file rather than the hardcoded time.

Also, it would be good to do this for a bunch of files at a time.

```{r message=FALSE}
library("tidyverse")
library("jsonlite")
source("../utils/hertzToST.R")

options(digits=6)
options(scipen=999)
options(tibble.print_max = Inf)

# IMPORTANT! Need /_voks folder inside whichever directory we are trying to run this
# To run with reprosody and reprosody_all.sh in ../other_scripts/

# Gesture files are in two places. Uncomment one at a time and run thhe whole thing
#path="../data/20_12_15-pilot_gestures/modalities/gestures/"
path="../data/21_05-fallrise/gestures/"
ref_freq = 116.54
getwd()
setwd(path)
filenames <- dir(pattern="\\.gest$")
setwd("../../../notebooks/")
getwd()

out<-sapply(filenames, gestToReprosody)
#out <- gestToReprosody(filenames[1])

gestToReprosody <- function(name) {
  # Read JSON into tibble
  data <- fromJSON(paste0(path, name)) %>% as_tibble()
  
  # Modify the data
  data <- data %>%
  # Convert t_out from millis to seconds
  mutate(t_out=round(t_init/1000, 4)) %>%
  # Keep t_in is scrub
  mutate(t_in=round(scrub, 2)) %>%
  # Change frequency to Hz
  mutate(f_hertz=round(semitones_to_hertz(f, ref_freq), 2)) %>%
  # Compute difference between t_out at position n & position n-1  
  mutate(t_out_diff = t_out - lag(t_out)) %>%
  # Remove any line where t_out_diff <= 0.005
  filter(t_out_diff > 0.005) %>%
  # Only keep the columns we need
  select(t_out, t_in, f_hertz)

  # Prepare to write output file
  # Get the part of the name before .gest
  # "." is special character, so we use \\.
  core_name <- str_split(name, "\\.")[[1]][1]
  filename_out <- paste0(path, "_voks/", core_name, ".txt")

  # Write the output
  cat("SCRUB", "\n", file=filename_out)
  write.table(data, filename_out, append=TRUE, col.names=FALSE, row.names=FALSE)
}



```
