==> notebooks/18_compareGestFallrise.Rmd <==
---
title: "Comparisons for Fallrise Data"
output: html_notebook
---
# Contents
0. Global definitions
1. File Loading:
  1.1 Reference recordings: DONE
    - PitchTier
    - TextGrids: because we want to specify the Region of Interest (ROI),
                 the tone-carrying syllables at the end of the sentence
  1.2 Gesture files
  1.3 Vocal files: 
Note: there are 2 versions of TextGrids for every vocal and gestural file, one with the ending consonant and one without.
  1.4: Info tibbles for all files

2. Preparing for comparison:
    - From TextGrids, identify the start and end time of ROI for each file
  2.1 Reference recordings
  2.2 Subjects' audio:
    2.2.1 Normalizing audio
    2.2.2 ROIs

Comparisons & score plots moved to 18b

3. Frequency plots

---------------------------
# 0. Global definitions
```{r}
library(tidyverse)
library(rPraat)
library(gridExtra)
library(jsonlite)
library(timetk)
library(zoo)
library(MetricsWeighted)
library(wesanderson)
source("../utils/hertzToST.R")
options(dplyr.print_max = 1e9)
options(digits=8)

ref_freq <- 116.54

# Loading PitchTier, interpolate values every 0.01 seconds
load_pitchtier <- function(filename) {
  my_pt <- pt.read(filename)
  interpolated <- pt.interpolate(my_pt, seq(0, my_pt$tmax, by=0.01))
  interpolated$f <- hertz_to_semitones(interpolated$f, ref_freq)
  return(interpolated)
}

load_intensitytier <- function(filename) {
  my_it <- it.read(filename)
  interpolated <- it.interpolate(my_it, seq(0, my_it$tmax, by=0.01))
  return(interpolated)
}

load_textgrid <- function(filename) {
  return(tg.read(filename))
}

load_gest <- function(filename) {
  #print(filename)
  fromJSON(filename) %>% 
  # Make it into a tibble, convert t_init & t_end to from milliseconds to seconds
  as_tibble() %>% mutate(t_init=t_init/1000, t_end=t_end/1000)  
}

ref_ids <- c('3a', '3b', '5a', '5b', 'Aa', 'Ab', 'Ca', 'Cb')
ref_ROI <- c("Credit (fall)", "Credit (fallrise)",
             "Good (fall)", "Good (fallrise)",
             "Cat (fall)", "Cat (fallrise)", 
             "Anyone (fall)", "Anyone (fallrise)")

# Get index of reference file
get_ref_index <- function(pt_filename) {
  phrase_id <- str_split(pt_filename, '-')[[1]][3]
  print(paste("Phrase Id:", phrase_id))
  ref_index <- match(phrase_id, ref_ids)
  return(ref_index)
}

# Return index of reference file if same=TRUE or unspecified
# Returns index of reference file with non-target intonation if same is FALSE
#    For instance, if the target is file Xa, it would return index of Xb
get_ref_index2 <- function(pt_filename, same=TRUE) {
  phrase_id <- str_split(pt_filename, '-')[[1]][3]
  phrase_letter <- substr(phrase_id, 2, 2)
  ref_index <- match(phrase_id, ref_ids)
  if (same) {
    return(ref_index)
  } else {
    if (phrase_letter=='a')
      return(ref_index+1)
    else
      return(ref_index-1)
  }
}

# Get the reference data based on phrase_id
get_ref_data <- function(ref_pts_roi, phrase_id) {
  index<-which(ref_ids==phrase_id)
  ref_data<-ref_pts_roi[[index]]
  num_points<-length(ref_data$f)
  my_data<-tibble(f=ref_data$f, t=ref_data$t, percent=seq(1, num_points)/num_points, condition="ref")
  return(my_data)
}



```


# 1. File Loading
## 1.1 Loading reference (demo) recordings
```{r}
# For REFERENCE recordings:
ref_path <- "../data/21_05-fallrise/reference/"

# - Load PitchTier and TextGrid filesnames
demo_pt_filenames <- dir(paste0(ref_path, "pt"))
demo_tg_filenames <- dir(paste0(ref_path, "tg"))
demo_tg2_filenames <- dir(paste0(ref_path, "tg_noend")) #no end
demo_it_filenames <- dir(paste0(ref_path, "it"))
# Use filenames to load PitchTier and TextGrid objects, stored in a list
ref_pts <- lapply(paste0(ref_path, "pt/",demo_pt_filenames), load_pitchtier)
ref_tgs <- lapply(paste0(ref_path, "tg/",demo_tg_filenames), load_textgrid)
ref_tgs2 <- lapply(paste0(ref_path, "tg_noend/",demo_tg2_filenames), load_textgrid)
ref_its <- lapply(paste0(ref_path, "it/",demo_it_filenames), load_intensitytier)

```

## 1.2 Loading gestures recorded from subjects
```{r}
# Load pitch contour (.gest) as JSON file
gest_path_learners <- "../data/21_05-fallrise/gest_learners/"
gest_path_natives <- "../data/21_05-fallrise/gest_natives/"

# Load filenames for gest & textgrids for LEARNERS
learners_gest_filenames <- dir(paste0(gest_path_learners, "gest"))
learners_tg_filenames <- dir(paste0(gest_path_learners, "tg"))
learners_tg2_filenames <- dir(paste0(gest_path_learners, "tg_noend"))
# Use filenames to load actual gest (as JSON) and textgrids
learners_gests <- lapply(paste0(gest_path_learners, "gest/", learners_gest_filenames), load_gest)
learners_gests_tgs <- lapply(paste0(gest_path_learners, "tg/", learners_tg_filenames), load_textgrid)
learners_gests_tgs2 <- lapply(paste0(gest_path_learners, "tg_noend/", learners_tg2_filenames), load_textgrid)

# Load filenames for gest & textgrids for NATIVES
natives_gest_filenames <- dir(paste0(gest_path_natives, "gest"))
natives_tg_filenames <- dir(paste0(gest_path_natives, "tg"))
natives_tg2_filenames <- dir(paste0(gest_path_natives, "tg_noend"))
# Use filenames to load actual gest (as JSON) and textgrids
natives_gests <- lapply(paste0(gest_path_natives, "gest/", natives_gest_filenames), load_gest)
natives_gests_tgs <- lapply(paste0(gest_path_natives, "tg/", natives_tg_filenames), load_textgrid)
natives_gests_tgs2 <- lapply(paste0(gest_path_natives, "tg_noend/", natives_tg2_filenames), load_textgrid)

# Both together
all_gests <- append(learners_gests, natives_gests)
all_gests_tgs <- append(learners_gests_tgs, natives_gests_tgs)
all_gests_tgs2 <- append(learners_gests_tgs2, natives_gests_tgs2)
```

## 1.3 Loading vocal recordings
```{r}
# For Vocal Recordings
voice_learners_path <- "../data/21_05-fallrise/voice_learners/"
voice_natives_path <- "../data/21_05-fallrise/voice_natives/"
  
# Load filenames of pitchtiers & textgrids for LEARNERS free interpretation
learners_pt_filenames <- dir(paste0(voice_learners_path, "pt"))
learners_tg_filenames <- dir(paste0(voice_learners_path, "tg"))
learners_tg2_filenames <- dir(paste0(voice_learners_path, "tg_noend"))
# Use filenames to load actual pitchtiers and textgrids
learners_pts <- lapply(paste0(voice_learners_path, "pt/", learners_pt_filenames), load_pitchtier)
learners_tgs <- lapply(paste0(voice_learners_path, "tg/", learners_tg_filenames), load_textgrid)
learners_tgs2 <- lapply(paste0(voice_learners_path, "tg_noend/", learners_tg2_filenames), load_textgrid)

# Load filenames for NATIVES, currently only 2 subjects free reading
natives_pt_filenames <- dir(paste0(voice_natives_path, "pt"))
natives_tg_filenames <- dir(paste0(voice_natives_path, "tg"))
natives_tg2_filenames <- dir(paste0(voice_natives_path, "tg_noend"))
# Use filenames to load actual pitchtiers and textgrids
natives_pts <- lapply(paste0(voice_natives_path, "pt/", natives_pt_filenames), load_pitchtier)
natives_tgs <- lapply(paste0(voice_natives_path, "tg/", natives_tg_filenames), load_textgrid)
natives_tgs2 <- lapply(paste0(voice_natives_path, "tg_noend/", natives_tg2_filenames), load_textgrid)

# Both together
all_pts <- append(learners_pts, natives_pts)
all_tgs <- append(learners_tgs, natives_tgs)
all_tgs2 <- append(learners_tgs2, natives_tgs2)
```


## 1.4 Info tibbles for all files
```{r}
learners_voice_info <- do.call("rbind", lapply(learners_pt_filenames, function(item) {
  name <- str_split(item, '.PitchTier')[[1]][1]
  info <- str_split(name, '-')[[1]]
  return(tibble(subject=info[1], type=info[2], pid=info[3], order=info[4], native=FALSE))
}))

natives_voice_info <- do.call("rbind", lapply(natives_pt_filenames, function(item) {
  name <- str_split(item, '.PitchTier')[[1]][1]
  info <- str_split(name, '-')[[1]]
  return(tibble(subject=info[1], type=info[2], pid=info[3], order=info[4], native=TRUE))
}))


learners_gest_info <- do.call("rbind", lapply(learners_gest_filenames, function(item) {
  name <- str_split(item, '.gest')[[1]][1]
  info <- str_split(name, '-')[[1]]
  return(tibble(subject=info[1], type=info[2], pid=info[3], order=info[4], native=FALSE))
}))

natives_gest_info <- do.call("rbind", lapply(natives_gest_filenames, function(item) {
  name <- str_split(item, '.gest')[[1]][1]
  info <- str_split(name, '-')[[1]]
  return(tibble(subject=info[1], type=info[2], pid=info[3], order=info[4], native=TRUE))
}))

# Combine learners and natives
voice_info <- full_join(learners_voice_info, natives_voice_info)
# Add column that encodes the index, to easily select pt/gest files
voice_info$index <- 1:nrow(voice_info)

gest_info <- full_join(learners_gest_info, natives_gest_info)
gest_info$index <- 1:nrow(gest_info)
```


# 2: Preparing for Comparison: 
## 2.1 Reference recordings
```{r}
# Generates a function to cut the pitchtiers in pts to the time boundaries of ROI 
# specified by the TextGrids in tgs
get_ref_pts_roi_prep <- function(tgs, pts) {
  function(index) {
    my_tg <- tgs[[index]]
    roi_start <- my_tg$study$t1[2]
    roi_end <- my_tg$study$t1[3]
    return(pt.cut(pts[[index]], tStart=roi_start, tEnd=roi_end))  
  }
}
# Same functiong generator to cut the intensity tiers
get_ref_its_roi_prep <- function(tgs, its) {
  function(index) {
    my_tg <- tgs[[index]]
    roi_start <- my_tg$study$t1[2]
    roi_end <- my_tg$study$t1[3]
    my_it<-it.cut(its[[index]], tStart=roi_start, tEnd=roi_end)
    # scale so that all the elements in the ROI add up to 1
    my_it$i<-my_it$i / sum(my_it$i)
    return(my_it)
  }
}

# Use the above functions: PTS
get_ref_pts_roi <- get_ref_pts_roi_prep(ref_tgs, ref_pts) 
get_ref_pts_roi2 <- get_ref_pts_roi_prep(ref_tgs2, ref_pts)

# For every PitchTier, only get the values in the ROI
ref_pts_roi <- seq(1, length(ref_tgs)) %>% lapply(get_ref_pts_roi)
ref_pts_roi2 <- seq(1, length(ref_tgs2)) %>% lapply(get_ref_pts_roi2)

# Use the above functions: ITS
get_ref_its_roi <- get_ref_its_roi_prep(ref_tgs, ref_its) 
get_ref_its_roi2 <- get_ref_its_roi_prep(ref_tgs2, ref_its) 

# For every IntensityTier, only get the values in the ROI, normalize it
ref_its_roi <- seq(1, length(ref_tgs)) %>% lapply(get_ref_its_roi)
ref_its_roi2 <- seq(1, length(ref_tgs)) %>% lapply(get_ref_its_roi2)
```

## 2.2 Preparing vocal recordings
### 2.2.1 Normalizing
```{r}
# Get the average pitch of every subject

# Helper function used by get_subject_mean_freq
get_freqs <- function(index) {return(tibble(f=all_pts[[index]]$f))}

generate_subject_mean_freq <- function(my_subject) {
  # Find all the indices of the pitchtiers for the subject
  pt_indices <- voice_info %>% filter(subject==my_subject) %>% select(index)  
  all_freqs <- do.call("rbind", lapply(pt_indices$index, get_freqs))  
  mean(all_freqs$f)
}

subjects <- c("sa2", "sa3", "sa4", "sa5", "sa6", "sa7", "sa8", "sa10", "sa11", "sa12", "sa13", "sa14",
              "sn1", "sn2", "sn3", "sn4", "sn5", "sn6", "sn7", "sn8", "sn9", "sn10", "sn11", "sn12")
means<- lapply(subjects, generate_subject_mean_freq) %>% unlist(recursive = FALSE, use.names = FALSE)
subjects_means <- tibble(subject=subjects, mean=means) 

get_subject_mean_freq <- function(my_subject) {
  info<-filter(subjects_means, subject==my_subject)
  return(info$mean)
}
```


### 2.2.2 ROIs for vocal recordings
```{r}
# returns a function that prepares subject pitch tier file
# specify vectors for the filenames, pitchtiers (in semitones), and textgrids 
# returns the pitch tier cut to region of interest, scaled to same  number of points as
# its reference
get_subject_pt_prep_function <- function(filenames, pts, tgs, ref_pts_roi, same=TRUE) {  
  function(index) {  
    # Print index for debugging: 
    #print(paste("Current Index:", index))
    item <- filenames[[index]]
    
    # Get the subject from filename
    filename<-str_split(item, '[.]')[[1]][1]
    # Get the parts of the filename
    filename_parts<-str_split(item, "-")[[1]]
    my_subject<- filename_parts[1]
    mean_freq <- get_subject_mean_freq(my_subject)
    
    # Find the index of that ref phrase
    ref_index <- get_ref_index2(item, same)
    # Get the ref pitch tier (ROI)
    ref_pt <- ref_pts_roi[[ref_index]]
    # Get the ROI of subject's vocal recording
    subject_roi_start <- tgs[[index]]$study$t1[2]
    subject_roi_end <- tgs[[index]]$study$t1[3]
    pt_roi <- pt.cut(pts[[index]], tStart=subject_roi_start, tEnd=subject_roi_end)
  
    # Interpolate based on the number of points in the ROI of the reference pitchtier
    pt_roi_interp <- pt.interpolate(pt_roi, 
                     seq(subject_roi_start, subject_roi_end, length=length(ref_pt$f)))
    pt_roi_interp$f <- pt_roi_interp$f-mean_freq
    return(pt_roi_interp)
  }
}

# Function to prepare the free vocal files for LEARNERS, then using it
learners_prep <- get_subject_pt_prep_function(learners_pt_filenames, learners_pts, learners_tgs, ref_pts_roi)
learners_pts_roi <- seq(1, length(learners_pt_filenames)) %>% lapply(learners_prep)
# For tgs2
learners_prep2 <- get_subject_pt_prep_function(learners_pt_filenames, learners_pts, learners_tgs2, ref_pts_roi2)
learners_pts_roi2 <- seq(1, length(learners_pt_filenames)) %>% lapply(learners_prep2)


# For comparison with non-target reference
learners_prep_nontarget <- get_subject_pt_prep_function(learners_pt_filenames, 
                                                        learners_pts, learners_tgs, ref_pts_roi, FALSE)
learners_pts_roi_nontarget <- seq(1, length(learners_pt_filenames)) %>% lapply(learners_prep_nontarget)


# Prepare free vocal files for NATIVES
natives_prep <- get_subject_pt_prep_function(natives_pt_filenames, natives_pts, natives_tgs, ref_pts_roi)
natives_pts_roi <- seq(1, length(natives_pt_filenames)) %>% lapply(natives_prep)
# For tgs2
natives_prep2 <- get_subject_pt_prep_function(natives_pt_filenames, natives_pts, natives_tgs2, ref_pts_roi2)
natives_pts_roi2 <- seq(1, length(natives_pt_filenames)) %>% lapply(natives_prep2)

# For comparison with non-target reference
natives_prep_nontarget <- get_subject_pt_prep_function(natives_pt_filenames, 
                                                       natives_pts, natives_tgs, ref_pts_roi, FALSE)
natives_pts_roi_nontarget <- seq(1, length(natives_pt_filenames)) %>% lapply(natives_prep_nontarget)

```

## 2.3 Preparing gestures
```{r}

get_subject_gest_prep_function <- function(filenames, gests, tgs, ref_pts_roi, same=TRUE) {
  function(index) {
    file <- filenames[[index]]
    # Print index for debugging: 
    print(paste("Current Index:", index))
    ref_index <- get_ref_index2(file, same)
    # Get ref pitch tier (ROI)
    ref_pt <- ref_pts_roi[[ref_index]]
    # Get the ROI of subjects gesture recording
    subject_roi_start <- tgs[[index]]$study$t1[2]
    subject_roi_end <- tgs[[index]]$study$t1[3]
    ## Get the gesture, only keep the ROI
    gest <- gests[[index]] %>% 
      filter(t_init >= subject_roi_start & t_init <= subject_roi_end) %>%
      rename(t = t_init) %>% select(t, f, scrub) %>%
      # filter out rows with duplicates in t
      filter(duplicated(t)==FALSE)
    
    # Add end points at subject_roi_start & subject_roi_end 
    # if they don't already exist, duplicating first and last available f value
    if (! subject_roi_start %in% gest$t) {
      gest <- add_row(gest, t=subject_roi_start, f=gest$f[1], scrub=gest$scrub[1], .before=1)
    }
    if (! subject_roi_end %in% gest$t) {
      gest <- add_row(gest, t=subject_roi_end, f=tail(gest$f, 1), scrub=tail(gest$scrub, 1))
    }
    
    # Interpolation magic - 
    # Goal: scale the subject gest ROI to have the same number of points
    # as the reference ROI
    num_samples <- length(ref_pt$f) 
    # Create a tibble with all the points we are interested in, with NA values
    sample_points <- tibble(t =seq(subject_roi_start, subject_roi_end, 
                                 length.out=num_samples), f=NA, scrub=NA) 

    # Add all the sample points whose scrub value doesn't already exist in data
    gest2 <- bind_rows(gest, filter(sample_points, !(t %in% gest$t))) %>% arrange(t)
      
    # Transform into zoo object and fill in NAs with interpolated values
    z <- read.zoo(gest2) %>% na.approx %>%
      tk_tbl(preserve_index=TRUE, rename_index="t") %>%
      filter(t %in% sample_points$t) 
    return(z)
  }
}

# Function to prepare the gesture files for LEARNERS, then using it to actually prep gestures
learners_gests_prep <- get_subject_gest_prep_function(learners_gest_filenames, 
                                                      learners_gests, learners_gests_tgs, ref_pts_roi)
learners_gests_roi <- seq(1, length(learners_gest_filenames)) %>% 
                      lapply(learners_gests_prep)
# For tgs2
learners_gests_prep2 <- get_subject_gest_prep_function(learners_gest_filenames, 
                                                      learners_gests, learners_gests_tgs2, ref_pts_roi2)
learners_gests_roi2 <- seq(1, length(learners_gest_filenames)) %>% 
                      lapply(learners_gests_prep2)


# Function to prepare gestures for LEARNERS with non-target reference phrase, then using it to prep gestures
learners_gests_prep_nontarget <- get_subject_gest_prep_function(learners_gest_filenames, 
                                                      learners_gests, learners_gests_tgs, ref_pts_roi, FALSE)
learners_gests_roi_nontarget <- seq(1, length(learners_gest_filenames)) %>%
                                lapply(learners_gests_prep_nontarget)

# Function to prepare the gesture files for NATIVES, then using it to actually prep gestures
natives_gests_prep <- get_subject_gest_prep_function(natives_gest_filenames, 
                                             natives_gests, natives_gests_tgs, ref_pts_roi)
natives_gests_roi <- seq(1, length(natives_gest_filenames)) %>% lapply(natives_gests_prep)

# For tgs2
natives_gests_prep2 <- get_subject_gest_prep_function(natives_gest_filenames, 
                                             natives_gests, natives_gests_tgs2, ref_pts_roi2)
natives_gests_roi2 <- seq(1, length(natives_gest_filenames)) %>% lapply(natives_gests_prep2)

# Function to prepare the gesture files for NATIVES using non-target reference phrase, then using it to actually prep gestures
natives_gests_prep_nontarget <- get_subject_gest_prep_function(natives_gest_filenames, 
                                             natives_gests, natives_gests_tgs, ref_pts_roi, FALSE)
natives_gests_roi_nontarget <- seq(1, length(natives_gest_filenames)) %>% lapply(natives_gests_prep_nontarget)
```


## 3. Frequency Plots
### 3.1. Preparing the data
```{r}

# NOTE!! Consider incorporating perception test results at this stage to make things go faster

# Function factory to associate information with frequency curves
get_curve_info <- function(filenames, data, nativeness) {
  function(index) {
    # Get the filename, remove extension
    filename<-filenames[[index]]
    #print(filename)
    filename<-str_split(filename, '[.]')[[1]][1]
    # Get the parts of the filename
    filename_parts<-str_split(filename, "-")[[1]]
    # Get the gesture data
    gest_data<-data[[index]] 
    # Get the number of points, to calculate "percent" time
    num_points<-length(gest_data$f)
    my_percent<-seq(1, num_points)/num_points
    # Add marker for start of the "event" (for autocorrelation stuff in GAMM)
    my_event<-rep(FALSE, num_points)
    my_event[1]<-TRUE
    my_data<-tibble(f=gest_data$f, t=gest_data$t, t.norm=my_percent,
                         subject=filename_parts[1], condition=filename_parts[2], 
                         phrase=filename_parts[3], fallrise=substring(filename_parts[[3]], 2, 2)=='b',
                         order=filename_parts[4], event=my_event, native=nativeness)
    return(my_data)
  }
}

# Function to get information for learners' gestures
get_learner_gest_info <- get_curve_info(learners_gest_filenames, learners_gests_roi, FALSE)
# All the learners gest data in a giant tibble
learners_gest_data<-do.call("rbind", lapply(seq(1, length(learners_gest_filenames)), get_learner_gest_info))
# tg2:
get_learner_gest_info2 <- get_curve_info(learners_gest_filenames, learners_gests_roi2, FALSE)
learners_gest_data2 <- do.call("rbind", lapply(seq(1, length(learners_gest_filenames)), get_learner_gest_info2))
# Entire curve, not only ROI
get_learner_gest_info_all <- get_curve_info(learners_gest_filenames, learners_gests, FALSE)
learners_gest_data_all<-do.call("rbind", lapply(seq(1, length(learners_gest_filenames)), get_learner_gest_info_all))

# Function to get information for natives' gestures
get_native_gest_info <- get_curve_info(natives_gest_filenames, natives_gests_roi, TRUE)
# All the native gest data in a giant tibble
natives_gest_data<-do.call("rbind", lapply(seq(1, length(natives_gest_filenames)), get_native_gest_info))
#tg2:
get_native_gest_info2 <- get_curve_info(natives_gest_filenames, natives_gests_roi2, TRUE)
natives_gest_data2<-do.call("rbind", lapply(seq(1, length(natives_gest_filenames)), get_native_gest_info2))
# Entire curve, not only ROI
get_native_gest_info_all <- get_curve_info(natives_gest_filenames, learners_gests, TRUE)
natives_gest_data_all<-do.call("rbind", lapply(seq(1, length(natives_gest_filenames)), get_native_gest_info_all))

# Function to get information for learners' voice data
get_learner_voice_info <- get_curve_info(learners_pt_filenames, learners_pts_roi, FALSE)
learners_voice_data<-do.call("rbind", lapply(seq(1, length(learners_pt_filenames)), get_learner_voice_info))
learners_voice_imitation_data <- filter(learners_voice_data, condition=="imitation")
learners_voice_reading_data <- filter(learners_voice_data, condition=="free")
#tg2
get_learner_voice_info2 <- get_curve_info(learners_pt_filenames, learners_pts_roi2, FALSE)
learners_voice_data2<-do.call("rbind", lapply(seq(1, length(learners_pt_filenames)), get_learner_voice_info2))
learners_voice_imitation_data2 <- filter(learners_voice_data, condition=="imitation")
learners_voice_reading_data2 <- filter(learners_voice_data, condition=="free")
# Entire curve, not only ROIs
get_learner_voice_info_all <- get_curve_info(learners_pt_filenames, learners_pts, FALSE)
learners_voice_data_all<-do.call("rbind", lapply(seq(1, length(learners_pt_filenames)), get_learner_voice_info_all))
learners_voice_imitation_data_all <- filter(learners_voice_data_all, condition=="imitation")
learners_voice_reading_data_all <- filter(learners_voice_data_all, condition=="free")


# Function to get information for natives voice info
get_native_voice_info <- get_curve_info(natives_pt_filenames, natives_pts_roi, TRUE)
natives_voice_data<-do.call("rbind", lapply(seq(1, length(natives_pt_filenames)), get_native_voice_info))
natives_voice_imitation_data <- filter(natives_voice_data, condition=="imitation")
natives_voice_reading_data <- filter(natives_voice_data, condition=="free")
#tg2
get_native_voice_info2 <- get_curve_info(natives_pt_filenames, natives_pts_roi2, TRUE)
natives_voice_data2<-do.call("rbind", lapply(seq(1, length(natives_pt_filenames)), get_native_voice_info2))
natives_voice_imitation_data2 <- filter(natives_voice_data, condition=="imitation")
natives_voice_reading_data2 <- filter(natives_voice_data, condition=="free")
# Entire curve, not only ROIS
get_native_voice_info_all <- get_curve_info(natives_pt_filenames, natives_pts, TRUE)
natives_voice_data_all<-do.call("rbind", lapply(seq(1, length(natives_pt_filenames)), get_native_voice_info_all))
natives_voice_imitation_data_all <- filter(natives_voice_data_all, condition=="imitation")
natives_voice_reading_data_all <- filter(natives_voice_data_all, condition=="free")

```


### 5.2 Plots
```{r}
library(cowplot)

# Plots the refefence gesture with all the subject gestures
plot_curves_with_ref <- function(f0_data, phrase_id, subject_group_name, condition, ref_pts_roi, note="") {
  ref_data<-get_ref_data(ref_pts_roi, phrase_id)
  my_f0_data <- f0_data %>% filter(phrase==phrase_id)
  ggplot(ref_data, mapping=aes(x=percent, y=f, color="Ref")) + 
  geom_line(size=1.5) +
  geom_line(data=my_f0_data, mapping=aes(x=t.norm, y=f, color=subject)) +
  labs(title=paste0("Aligned Ref and Subject F0s, ", subject_group_name, " ", note),
                 subtitle=paste0("Phrase: ", phrase_id, ",Condition: ", condition), 
                 y="Frequency (Semitones)", x="Percent time")
}


# Function that makes and saves plots for diff data sets
save_curves_ref_plots <- function(f0_data, subject_group, condition, ref_pts_roi, note="") {
  plot3a <- plot_curves_with_ref(f0_data, "3a", subject_group, condition, ref_pts_roi)
  plot3b <- plot_curves_with_ref(f0_data, "3b", subject_group, condition, ref_pts_roi)
  plot5a <- plot_curves_with_ref(f0_data, "5a", subject_group, condition, ref_pts_roi)
  plot5b <- plot_curves_with_ref(f0_data, "5b", subject_group, condition, ref_pts_roi)
  plotAa <- plot_curves_with_ref(f0_data, "Aa", subject_group, condition, ref_pts_roi)
  plotAb <- plot_curves_with_ref(f0_data, "Ab", subject_group, condition, ref_pts_roi)
  plotCa <- plot_curves_with_ref(f0_data, "Ca", subject_group, condition, ref_pts_roi)
  plotCb <- plot_curves_with_ref(f0_data, "Cb", subject_group, condition, ref_pts_roi)
  curves35 <- plot_grid(plot3a, plot3b, plot5a, plot5b, ncol=2)
  curvesAC <- plot_grid(plotAa, plotAb, plotCa, plotCb, ncol=2)
  filename35 <- paste("3ab5ab", subject_group, condition, note, sep="_")
  filenameAC <- paste("AabCab", subject_group, condition, note, sep="_")
  ggsave(paste0(filename35, ".png"), plot=curves35, device='png', path="../data/21_05-fallrise/_plots", width = 10, height = 8)
  ggsave(paste0(filenameAC, ".png"), plot=curvesAC, device='png', path="../data/21_05-fallrise/_plots", width = 10, height = 8)
}

save_curves_ref_plots(learners_gest_data, "Learners", "GesturalImitation", ref_pts_roi)
save_curves_ref_plots(natives_gest_data, "Natives", "GesturalImitation", ref_pts_roi)
save_curves_ref_plots(learners_voice_imitation_data, "Learners", "VocalImitation", ref_pts_roi)
save_curves_ref_plots(natives_voice_imitation_data, "Natives", "VocalImitation", ref_pts_roi)

save_curves_ref_plots(learners_gest_data2, "Learners", "GesturalImitation", ref_pts_roi2, "ne")
save_curves_ref_plots(natives_gest_data2, "Natives", "GesturalImitation", ref_pts_roi2, "ne")
save_curves_ref_plots(learners_voice_imitation_data2, "Learners", "VocalImitation", ref_pts_roi2, "ne")
save_curves_ref_plots(natives_voice_imitation_data2, "Natives", "VocalImitation", ref_pts_roi2, "ne")
save_curves_ref_plots(learners_voice_imitation_data2, "Learners", "VocalReading", ref_pts_roi2, "ne")
save_curves_ref_plots(natives_voice_imitation_data2, "Natives", "VocalReading", ref_pts_roi2, "ne")


  # All
save_curves_ref_plots(learners_voice_reading_data_all, "Learners", "VocalReading", ref_pts, "All")
save_curves_ref_plots(natives_voice_reading_data_all, "Natives", "VocalReading", ref_pts, "All")
save_curves_ref_plots(learners_voice_imitation_data_all, "Learners", "VocalImitation", ref_pts, "All")
save_curves_ref_plots(natives_voice_imitation_data_all, "Natives", "VocaImitation", ref_pts, "All")
save_curves_ref_plots(learners_gest_data_all, "Learners", "GesturalImitation", ref_pts, "All")
save_curves_ref_plots(natives_gest_data_all, "Natives", "GesturalImitation", ref_pts, "All")

```
